+++ packages/react/src/hooks/useConnect.test-d.ts
@@ -67,22 +67,26 @@ test('context', () => {
     | undefined
   >()
   expectTypeOf(error).toEqualTypeOf<ConnectErrorType | null>()
-  expectTypeOf(variables).toEqualTypeOf<
+  expectTypeOf(variables).toMatchTypeOf<
     | {
         chainId?: number | undefined
-        connector: Connector | CreateConnectorFn
+        connector: CreateConnectorFn | Connector
       }
     | undefined
   >()
   expectTypeOf(context).toEqualTypeOf<typeof contextValue | undefined>()
 
   connect(
-    { connector },
+    {
+      connector,
+      foo: 'bar',
+    },
     {
       onError(error, variables, context) {
         expectTypeOf(variables).toEqualTypeOf<{
           chainId?: number | undefined
-          connector: Connector | CreateConnectorFn
+          connector: typeof connector | CreateConnectorFn
+          foo?: string | undefined
         }>()
         expectTypeOf(error).toEqualTypeOf<ConnectErrorType>()
         expectTypeOf(context).toEqualTypeOf<typeof contextValue | undefined>()
@@ -90,7 +94,8 @@ test('context', () => {
       onSuccess(data, variables, context) {
         expectTypeOf(variables).toEqualTypeOf<{
           chainId?: number | undefined
-          connector: Connector | CreateConnectorFn
+          connector: typeof connector | CreateConnectorFn
+          foo?: string | undefined
         }>()
         expectTypeOf(data).toEqualTypeOf<{
           accounts: readonly [Address, ...Address[]]
@@ -109,7 +114,8 @@ test('context', () => {
         expectTypeOf(error).toEqualTypeOf<ConnectErrorType | null>()
         expectTypeOf(variables).toEqualTypeOf<{
           chainId?: number | undefined
-          connector: Connector | CreateConnectorFn
+          connector: typeof connector | CreateConnectorFn
+          foo?: string | undefined
         }>()
         expectTypeOf(context).toEqualTypeOf<typeof contextValue | undefined>()
       },