+++ packages/register-tests/react/src/config.ts
@@ -1,9 +1,10 @@
 import { http } from 'viem'
-import { createConfig } from 'wagmi'
+import { createConfig, mock } from 'wagmi'
 import { celo, mainnet, optimism, zkSync } from 'wagmi/chains'
 
 export const config = createConfig({
   chains: [celo, mainnet, optimism, zkSync],
+  connectors: [mock({ accounts: ['0x'] })],
   transports: {
     [celo.id]: http(),
     [mainnet.id]: http(),